+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Magento 2 Goomento Base
 
+## 1.0.1: Dynamic Instance Manager calling added
+### Features:
+
+Added:
+- Rate traits
+- Datetime traits
+
 ## 1.0.0: First release
 ### Features: